encoder/requirements.txt
boto3==1.37.18
botocore==1.37.18
click==8.1.8
jmespath==1.0.1
psycopg2-binary==2.9.10
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ffmpeg==2.0.12
s3transfer==0.11.4
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
